Jutul

Documentation for Jutul.

Jutul.AMGPreconditioner

Jutul.BlockMajorLayout

Jutul.CartesianMesh

Jutul.CompactAutoDiffCache

Jutul.ConstantVariables

Jutul.DampedJacobiPreconditioner

Jutul.DefaultContext

Jutul.EquationMajorLayout

Jutul.FlowDiscretization

Jutul.GroupWisePreconditioner

Jutul.ILUZeroPreconditioner

Jutul.JutulAutoDiffCache

Jutul.JutulDiscretization

Jutul.LUPreconditioner

Jutul.ParallelCSRContext

Jutul.SPU

Jutul.TPFA

Jutul.TwoPointFiniteVolumeGeometry

Jutul.UnaryTabulatedVariable

Jutul.UnitMajorLayout

Jutul.absolute_increment_limit

Jutul.align_to_jacobian!

Jutul.align_to_jacobian!

Jutul.allocate_array_ad

Jutul.allocate_array_ad

Jutul.allocate_array_ad

Jutul.apply_forces!

Jutul.apply_forces_to_equation!

Jutul.as_value

Jutul.associated_entity

Jutul.associated_entity

Jutul.cell_dims

Jutul.cell_index

Jutul.convert_state_ad

Jutul.coord_offset

Jutul.declare_pattern

Jutul.declare_sparsity

Jutul.degrees_of_freedom_per_entity

Jutul.degrees_of_freedom_per_entity

Jutul.dim

Jutul.full_cell

Jutul.get_ad_entity_scalar

Jutul.get_dependencies

Jutul.get_entity_tag

Jutul.get_entries

Jutul.get_entries

Jutul.get_primary_variable_ordered_entities

Jutul.global_cell

Jutul.global_face

Jutul.initialize_context!

Jutul.initialize_extra_state_fields!

Jutul.initialize_storage!

Jutul.initialize_variable_value!

Jutul.interior_cell

Jutul.local_ad

Jutul.local_cell

Jutul.local_face

Jutul.maximum_value

Jutul.minimum_value

Jutul.number_of_cells

Jutul.number_of_degrees_of_freedom

Jutul.number_of_entities

Jutul.number_of_entities

Jutul.number_of_entities

Jutul.number_of_entities

Jutul.number_of_entities

Jutul.number_of_equations

Jutul.number_of_equations_per_entity

Jutul.number_of_faces

Jutul.number_of_partials_per_entity

Jutul.read_results

Jutul.relative_increment_limit

Jutul.replace_variables!

Jutul.set_parameters!

Jutul.set_primary_variables!

Jutul.set_secondary_variables!

Jutul.setup_state

Jutul.setup_state!

Jutul.setup_storage

Jutul.setup_storage!

Jutul.spu_upwind

Jutul.spu_upwind_mult

Jutul.spu_upwind_mult_index

Jutul.synchronize

Jutul.tpfv_geometry

Jutul.transfer

Jutul.two_point_potential_drop

Jutul.two_point_potential_drop_half_face

Jutul.update_before_step!

Jutul.update_equation!

Jutul.update_linearized_system_equation!

Jutul.update_secondary_variable!

Jutul.update_state_dependents!

Jutul.update_values!

Jutul.value

Jutul.value

Jutul.values_per_entity

Jutul.@jutul_secondary

Jutul.AMGPreconditioner — Type

AMG on CPU (Julia native)

Jutul.BlockMajorLayout — Type

Same as UnitMajorLayout, but the nzval is a matrix

Jutul.CartesianMesh — Type

CartesianMesh(dims, [Δ, [origin]])

Create a Cartesian mesh with dimensions specified by the Tuple dims.

Arguments

dims::Tuple: Number of grid cells in each direction. For example, (nx, ny) will give a 2D grids with nx cells in the x-direction.

Δ::Tuple=ones(length(dims)): Equal length to dims. First option: A Tuple of scalars where each entry is the length of each cell in that direction. For

example, specifying (Δx, Δy) for a uniform grid with each grid cell having area ofΔx*Δy. Second option: ATuple` of vectors where each entry contains the cell sizes in the direction.

origin=zeros(length(dims)): The origin of the first corner in the grid.

Examples

Generate a uniform 3D mesh that discretizes a domain of 2 by 3 by 5 units with 3 by 5 by 2 cells:

julia> CartesianMesh((3, 5, 2), (2.0, 3.0, 5.0))
CartesianMesh (3D) with 3x5x2=30 cells

Generate a non-uniform 2D mesh:

julia> CartesianMesh((2, 3), ([1.0, 2.0], [0.1, 3.0, 2.5]))
CartesianMesh (3D) with 3x5x2=30 cells

Jutul.CompactAutoDiffCache — Type

Cache that holds an AD vector/matrix together with their positions.

Jutul.ConstantVariables — Type

A set of constants, repeated over the entire set of Cells or some other entity

Jutul.DampedJacobiPreconditioner — Type

Damped Jacobi preconditioner on CPU

Jutul.DefaultContext — Type

Default context

Jutul.EquationMajorLayout — Type

Equations are stored sequentially in rows, derivatives of same type in columns:

Jutul.FlowDiscretization — Type

Discretization of kgradp + upwind

Jutul.GroupWisePreconditioner — Type

Multi-model preconditioners

Jutul.ILUZeroPreconditioner — Type

ILU(0) preconditioner on CPU

Jutul.JutulAutoDiffCache — Type

An AutoDiffCache is a type that holds both a set of AD values and a map into some global Jacobian.

Jutul.JutulDiscretization — Type

Ask discretization for entry i for specific entity

Jutul.LUPreconditioner — Type

Full LU factorization as preconditioner (intended for smaller subsystems)

Jutul.ParallelCSRContext — Type

Context that uses threads etc to accelerate loops

Jutul.SPU — Type

Single-point upwinding.

Jutul.TPFA — Type

Two-point flux approximation.

Jutul.TwoPointFiniteVolumeGeometry — Type

TwoPointFiniteVolumeGeometry(neighbors, areas, volumes, normals, cell_centers, face_centers)

Store two-point geometry information for a given list of neighbors specified as a 2 by n matrix where n is the number of faces such that face i connectes cells N[1, i] and N[2, i].

The two-point finite-volume geometry contains the minimal set of geometry information required to compute standard finite-volume discretizations.

Jutul.UnaryTabulatedVariable — Type

Jutul.UnitMajorLayout — Type

Domain entities sequentially in rows:

Jutul.absolute_increment_limit — Method

Absolute allowable change for variable during a nonlinear update.

Jutul.align_to_jacobian! — Method

Update an equation so that it knows where to store its derivatives in the Jacobian representation.

Jutul.align_to_jacobian! — Method

Update positions of law's derivatives in global Jacobian

Jutul.allocate_array_ad — Method

allocate_array_ad(v::AbstractMatrix, ...)

Convert matrix to AD matrix.

Jutul.allocate_array_ad — Method

allocate_array_ad(v::AbstractVector, ...)

Convert vector to AD vector.

Jutul.allocate_array_ad — Method

allocate_array_ad(n[, m]; <keyword arguments>)

Allocate vector or matrix as AD with optionally provided context and a specified non-zero on the diagonal.

Arguments

n::Integer: number of entries in vector, or number of rows if m is given.

m::Integer: number of rows (optional)

Keyword arguments

npartials = 1: Number of partials derivatives to allocate for each element

diag_pos = nothing: Indices of where to put entities on the diagonal (if any)

Other keyword arguments are passed onto get_ad_entity_scalar.

Examples:

Allocate a vector with a single partial:

julia> allocate_array_ad(2)
2-element Vector{ForwardDiff.Dual{nothing, Float64, 1}}:
 Dual{nothing}(0.0,0.0)
 Dual{nothing}(0.0,0.0)

Allocate a vector with two partials, and set the first to one:

julia> allocate_array_ad(2, diag_pos = 1, npartials = 2)
2-element Vector{ForwardDiff.Dual{nothing, Float64, 2}}:
 Dual{nothing}(0.0,1.0,0.0)
 Dual{nothing}(0.0,1.0,0.0)

Set up a matrix with two partials, where the first column has partials [1, 0] and the second [0, 1]:

julia> allocate_array_ad(2, 2, diag_pos = [1, 2], npartials = 2)
2×2 Matrix{ForwardDiff.Dual{nothing, Float64, 2}}:
 Dual{nothing}(0.0,1.0,0.0)  Dual{nothing}(0.0,1.0,0.0)
 Dual{nothing}(0.0,0.0,1.0)  Dual{nothing}(0.0,0.0,1.0)

Jutul.apply_forces! — Method

Apply a set of forces to all equations. Equations that don't support a given force will just ignore them, thanks to the power of multiple dispatch.

Jutul.apply_forces_to_equation! — Method

Update an equation with the effect of a force. The default behavior for any force we do not know about is to assume that the force does not impact this particular equation.

Jutul.as_value — Method

Create a mapped array that produces only the values when indexed.

Only useful for AD arrays, otherwise it does nothing.

Jutul.associated_entity — Method

Return the domain entity the equation is associated with

Jutul.associated_entity — Method

The entity a variable is associated with, and can hold partial derivatives with respect to.

Jutul.cell_dims — Method

cell_dims(g, pos)::Tuple

Get physical cell dimensions of cell with index pos for grid g.

Jutul.cell_index — Method

cell_index(g, pos)

Get linear (scalar) index of mesh cell from provided IJK tuple pos.

Jutul.convert_state_ad — Function

Convert a state containing variables as arrays of doubles to a state where those arrays contain the same value as Dual types. The dual type is currently taken from ForwardDiff.

Jutul.coord_offset — Method

Lower corner for one dimension, without any transforms applied

Jutul.declare_pattern — Method

Give out source, target arrays of equal length for a given equation attached to the given model.

Jutul.declare_sparsity — Method

Give out I, J arrays of equal length for a given equation attached to the given model.

Jutul.degrees_of_freedom_per_entity — Method

Constant variables hold no degrees of freedom.

Jutul.degrees_of_freedom_per_entity — Method

Number of independent primary variables / degrees of freedom per computational entity.

Jutul.dim — Method

dim(g)::Integer

Get the dimension of a mesh.

Jutul.full_cell — Method

Inner cell to local cell (full set)

Jutul.get_ad_entity_scalar — Method

get_ad_entity_scalar(v::Real, npartials, diag_pos = nothing; <keyword_arguments>)

Get scalar with partial derivatives as AD instance.

Arguments

v::Real: Value of AD variable.

npartials: Number of partial derivatives each AD instance holds.

diag_pos = nothing: Position(s) of where to set 1 as the partial derivative instead of zero.

Keyword arguments

tag = nothing: Tag for AD instance. Two AD values of the different tag cannot interoperate to avoid perturbation confusion (see ForwardDiff documentation).

Jutul.get_dependencies — Method

Get dependencies of variable when viewed as a secondary variable. Normally autogenerated with @jutul_secondary

Jutul.get_entity_tag — Method

Combine a base tag (which can be nothing) with a entity to get a tag that captures base tag + entity tag for use with AD initialization.

Jutul.get_entries — Method

Get entries of autodiff cache. Entries are AD vectors that hold values and derivatives.

Jutul.get_entries — Method

Get the entries of the main autodiff cache for an equation.

Note: This only gets the .equation field's entries.

Jutul.get_primary_variable_ordered_entities — Method

Get only the entities where primary variables are present, sorted by their order in the primary variables.

Jutul.global_cell — Method

Local cell -> global cell (full set)

Jutul.global_face — Method

Local face -> global face (full set)

Jutul.initialize_context! — Method

Initialize context when setting up a model

Jutul.initialize_extra_state_fields! — Method

Add variables that need to be in state, but are never AD variables (e.g. phase status flag)

Jutul.initialize_storage! — Method

Initialize the already allocated storage at the beginning of a simulation. Use this to e.g. set up extra stuff in state0 needed for initializing the simulation loop.

Jutul.initialize_variable_value! — Method

Initializer for the value of non-scalar primary variables

Jutul.interior_cell — Method

Local cell in full set -> inner cell (or zero)

Jutul.local_ad — Method

local_ad(state::T, index::I, ad_tag::∂T) where {T, I<:Integer, ∂T}

Create localad for state for index I of AD tag of type adtag

Jutul.local_cell — Method

Global cell -> local cell (full set)

Jutul.local_face — Method

Global face -> local face (full set)

Jutul.maximum_value — Method

Upper (inclusive) limit for variable.

Jutul.minimum_value — Method

Lower (inclusive) limit for variable.

Jutul.number_of_cells — Method

number_of_cells(g)::Integer

Get the number of cells in a mesh.

Jutul.number_of_degrees_of_freedom — Method

Total number of degrees of freedom for a model, over all primary variables and all entities.

Jutul.number_of_entities — Method

Get the number of entities (e.g. the number of cells) that the equation is defined on.

Jutul.number_of_entities — Method

Number of entities (e.g. Cells, Faces) a variable is defined on. By default, each primary variable exists on all cells of a discretized domain

Jutul.number_of_entities — Method

Get number of entities a cache is defined on.

Jutul.number_of_entities — Method

Number of entities for vector stored in state (just the number of elements)

Jutul.number_of_entities — Method

Number of entities for matrix stored in state (convention is number of columns)

Jutul.number_of_equations — Method

Get the total number of equations on the domain of model.

Jutul.number_of_equations_per_entity — Method

Get the number of equations per entity. For example, mass balance of two components will have two equations per grid cell (= entity)

Jutul.number_of_faces — Method

number_of_faces(g)::Integer

Get the number of faces in a mesh.

Jutul.number_of_partials_per_entity — Method

Get the number of partials

Jutul.read_results — Method

Read results from a given outputpath provded to simulate or simulatorconfig

Jutul.relative_increment_limit — Method

Relative allowable change for variable during a nonlinear update. A variable with value |x| and relative limit 0.2 cannot change more than |x|*0.2.

Jutul.replace_variables! — Method

Replace a variable that already exists (either primary or secondary)

Jutul.set_parameters! — Method

Set a parameter (adding if it does not exist)

Jutul.set_primary_variables! — Method

Set a primary variable (adding if it does not exist)

Jutul.set_secondary_variables! — Method

Set a secondary variable (adding if it does not exist)

Jutul.setup_state! — Function

Initialize primary variables and other state fields, given initial values as a Dict

Jutul.setup_state — Method

Set up a state. You likely want to overload setup_state! instead of this one.

Jutul.setup_storage! — Method

Allocate storage for a given model. The storage consists of all dynamic quantities used in the simulation. The default implementation allocates properties, equations and linearized system.

Jutul.setup_storage — Method

Allocate storage for the model. You should overload setup_storage! if you have a custom definition.

Jutul.spu_upwind — Method

Perform single-point upwinding based on signed potential.

Jutul.spu_upwind_mult — Method

Perform single-point upwinding based on signed potential, then multiply the result with that potential

Jutul.spu_upwind_mult_index — Method

Perform single-point upwinding based on signed potential, then multiply the result with that potential

Jutul.synchronize — Method

Synchronize backend after allocations.

Some backends may require notification that storage has been allocated.

Jutul.tpfv_geometry — Function

tpfv_geometry(g)

Generate two-point finite-volume geometry for a given grid, if supported.

See also TwoPointFiniteVolumeGeometry.

Jutul.transfer — Method

Transfer v to the representation expected by a given context.

For the defalt context, the transfer function does nothing. For other context such as the CUDA version, it may convert integers and floats to other types (e.g. Float32) and Arrays to CuArrays.

You will likely have to implement some transfer operators for your own types if you want to simulate with a non-default context.

Jutul.two_point_potential_drop — Method

Two-point potential drop with gravity (generic)

Jutul.two_point_potential_drop_half_face — Method

Two-point potential drop (with derivatives only respect to "c_self")

Jutul.update_before_step! — Method

Jutul.update_equation! — Method

Update equation based on currently stored properties

Jutul.update_linearized_system_equation! — Method

Update a linearized system based on the values and derivatives in the equation.

Jutul.update_secondary_variable! — Method

Update a secondary variable. Normally autogenerated with @jutul_secondary

Jutul.update_state_dependents! — Method

Perform updates of everything that depends on the state.

This includes properties, governing equations and the linearized system

Jutul.update_values! — Method

update_values!(x, dx)

Replace values of x in-place by y, leaving x with the avlues of y and the partials of x.

Jutul.value — Method

value(d::Dict)

Call value on all elements of some Dict.

Jutul.value — Method

Take value of AD.

Jutul.values_per_entity — Method

Number of values held by a primary variable. Normally this is equal to the number of degrees of freedom, but some special primary variables are most conveniently defined by having N values and N-1 independent variables.

Jutul.@jutul_secondary — Macro

Designate the function as updating a secondary variable.

The function is then declared, in addition to helpers that allows checking what the dependencies are and unpacking the dependencies from state.

If we define the following function annotated with the macro: @jutulsecondary function updateas_secondary!(target, var::MyVarType, model, a, b, c) @. target = a + b / c end

The macro also defines: function get_dependencies(var::MyVarType, model) return [:a, :b, :c] end

function updatesecondaryvariable!(arraytarget, var::MyVarType, model, state) updateassecondary!(arraytarget, var, model, state.a, state.b, state.c) end

Note that the names input arguments beyond the parameter dict matter, as these will be fetched from state.